Fix SAT solver and SMT theories: correctness, null safety, type errors
sat_simplifier: remove duplicate m_elim_counter decrement (double-counting)
sat_solver: fix swapped log/exp in logsumexp activity calculation
arith_solver: fix swapped upper/lower bound assertion stats
array_internalize: fix select loop bound num_args-1 -> num_args
bv_delay_internalize: remove dead return, fix bit2bool swapped args (j,i)
bv_solver.h: zero-initialize bv_justification enode pointers
dt_solver: fix signed/unsigned mismatches, add null checks for array
  theory and constructor enode, add fresh value fallback for no-constructor
intblast_solver: use safe map lookup with find() for unsat core
pb_card: add missing Z3_fallthrough between l_true and l_undef cases
pb_solver: add SASSERT(c > 0) in divide to guard against division by zero
q_ematch: fix sizeof(enode* const*) -> sizeof(enode*) in two allocations
sat_th.h: fix is_true(a,b,c,d) checking c instead of d
smt_quantifier: fix lazy_scopes unsigned underflow in post-decrement,
  fix max_top_generation forwarded as max_generation
theory_arith_aux: add missing free-list head update in del_row_entry
theory_arith_core: fix rem/mod using zero instead of original divisor,
  fix numeral*numeral using addition instead of multiplication
theory_array_full: widen unsigned short loop counter to unsigned
theory_bv: fix reentrancy bug — use local vector instead of m_bits_expr
  class member that gets reset during recursive ctx.internalize calls
theory_lra: fix rem/mod using mk_int(0) instead of original operands
theory_seq: fix dependency leak by resetting dep before canonize,
  fix eq_id pre-increment vs post-increment consistency,
  fix solution_map reserve allocating 2x needed memory

+++ src/smt/theory_array_full.cpp
@@ -612,7 +612,7 @@ namespace smt {
         m_stats.m_num_select_const_axiom++;   
         ptr_buffer<expr> sel_args;
         sel_args.push_back(cnst->get_expr());
-        for (unsigned short i = 1; i < num_args; ++i) {
+        for (unsigned i = 1; i < num_args; ++i) {
             sel_args.push_back(select->get_expr()->get_arg(i));
         }
         expr * sel = mk_select(sel_args.size(), sel_args.data());
@@ -646,7 +646,7 @@ namespace smt {
         m_stats.m_num_select_as_array_axiom++;   
         ptr_buffer<expr> sel_args;
         sel_args.push_back(arr->get_expr());
-        for (unsigned short i = 1; i < num_args; ++i) {
+        for (unsigned i = 1; i < num_args; ++i) {
             sel_args.push_back(select->get_expr()->get_arg(i));
         }
         expr * sel = mk_select(sel_args.size(), sel_args.data());
